Trait jacs::agent::loaders::FileLoader

pub trait FileLoader {

    // Required methods
    fn build_filepath(
        &self,
        doctype: &String,
        docid: &String
    ) -> Result<PathBuf, Box<dyn Error>>;
    fn build_key_filepath(
        &self,
        doctype: &String,
        key_filename: &String
    ) -> Result<PathBuf, Box<dyn Error>>;
    fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Box<dyn Error>>;
    fn fs_agent_load(&self, agentid: &String) -> Result<String, Box<dyn Error>>;
    fn fs_agent_new(&self, filename: &String) -> Result<String, Box<dyn Error>>;
    fn fs_document_new(
        &self,
        filename: &String
    ) -> Result<String, Box<dyn Error>>;
    fn fs_document_load(
        &self,
        document_id: &String
    ) -> Result<String, Box<dyn Error>>;
    fn fs_preload_keys(
        &mut self,
        private_key_filename: &String,
        public_key_filename: &String
    ) -> Result<(), Box<dyn Error>>;
    fn fs_save_keys(&mut self) -> Result<(), Box<dyn Error>>;
    fn fs_load_keys(&mut self) -> Result<(), Box<dyn Error>>;
    fn fs_docs_save_all(&mut self) -> Result<Vec<String>, Box<dyn Error>>;
    fn fs_agent_save(
        &self,
        agentid: &String,
        agent_string: &String
    ) -> Result<String, Box<dyn Error>>;
    fn fs_document_save(
        &self,
        document_id: &String,
        document_string: &String
    ) -> Result<String, Box<dyn Error>>;

}

The goal of fileloader is to prevent fileloading into arbitrary directories by centralizing all filesystem access Only an initilaized agent can perform some of the functions by calling isready() as an attempt to ensure actions on the filesystem requiring the agent are acted out by the agent

Required Methods

fn build_filepath( &self, doctype: &String, docid: &String ) -> Result<PathBuf, Box<dyn Error>>

fn build_key_filepath( &self, doctype: &String, key_filename: &String ) -> Result<PathBuf, Box<dyn Error>>

fn fs_docs_load_all(&mut self) -> Result<Vec<String>, Box<dyn Error>>

fn fs_agent_load(&self, agentid: &String) -> Result<String, Box<dyn Error>>

fn fs_agent_new(&self, filename: &String) -> Result<String, Box<dyn Error>>

fn fs_document_new(&self, filename: &String) -> Result<String, Box<dyn Error>>

fn fs_document_load( &self, document_id: &String ) -> Result<String, Box<dyn Error>>

fn fs_preload_keys( &mut self, private_key_filename: &String, public_key_filename: &String ) -> Result<(), Box<dyn Error>>

fn fs_save_keys(&mut self) -> Result<(), Box<dyn Error>>

fn fs_load_keys(&mut self) -> Result<(), Box<dyn Error>>

fn fs_docs_save_all(&mut self) -> Result<Vec<String>, Box<dyn Error>>

fn fs_agent_save( &self, agentid: &String, agent_string: &String ) -> Result<String, Box<dyn Error>>

fn fs_document_save( &self, document_id: &String, document_string: &String ) -> Result<String, Box<dyn Error>>