Struct jacs::agent::Agent

pub struct Agent { /* private fields */ }

Implementations

impl Agent

pub fn new( agentversion: &String, headerversion: &String, signature_version: &String ) -> Result<Self, Box<dyn Error>>

pub fn load_by_id( &mut self, id: Option<String>, _version: Option<String> ) -> Result<(), Box<dyn Error>>

pub fn ready(&mut self) -> bool

pub fn set_keys( &mut self, private_key: Vec<u8>, public_key: Vec<u8>, key_algorithm: &String ) -> Result<(), Box<dyn Error>>

pub fn get_private_key(&self) -> Result<Secret<PrivateKey>, Box<dyn Error>>

pub fn get_default_dir(&self) -> PathBuf

pub fn load(&mut self, agent_string: &String) -> Result<(), Box<dyn Error>>

pub fn verify_self_signature(&mut self) -> Result<(), Box<dyn Error>>

pub fn get_agent_for_doc( &mut self, document_key: String, signature_key_from: Option<&String> ) -> Result<String, Box<dyn Error>>

pub fn verify_hash(&self, doc: &Value) -> Result<bool, Box<dyn Error>>

verify the hash of a complete document that has SHA256_FIELDNAME

pub fn verify_self_hash(&self) -> Result<bool, Box<dyn Error>>

verify the hash where the document is the agent itself.

pub fn get_schema_keys(&mut self) -> Vec<String>

pub fn update_self( &mut self, new_agent_string: &String ) -> Result<String, Box<dyn Error>>

pass in modified agent’s JSON the function will replace it’s internal value after: versioning resigning rehashing

pub fn validate_header( &mut self, json: &str ) -> Result<Value, Box<dyn Error + 'static>>

pub fn validate_agent( &mut self, json: &str ) -> Result<Value, Box<dyn Error + 'static>>

pub fn load_custom_schemas(&mut self, schema_paths: &[String])

pub fn save(&self) -> Result<String, Box<dyn Error>>

pub fn create_agent_and_load( &mut self, json: &String, create_keys: bool, _create_keys_algorithm: Option<&String> ) -> Result<Value, Box<dyn Error + 'static>>

create an agent, and provde id and version as a result